Set `get_preset` for Kusama/Polkadot + chain-spec-generator clean up + taplo
