+++ chain-spec-generator/Cargo.toml
@@ -12,20 +12,9 @@ serde_json = { workspace = true }
 serde = { features = ["derive"], workspace = true }
 
 polkadot-runtime = { workspace = true }
-polkadot-runtime-constants = { workspace = true, default-features = true }
 kusama-runtime = { workspace = true }
-kusama-runtime-constants = { workspace = true, default-features = true }
 
 sc-chain-spec = { workspace = true }
-runtime-parachains = { default-features = true, workspace = true }
-polkadot-primitives = { workspace = true, default-features = true }
-babe-primitives = { workspace = true, default-features = true }
-authority-discovery-primitives = { workspace = true, default-features = true }
-sp-core = { workspace = true, default-features = true }
-pallet-staking = { workspace = true, default-features = true }
-grandpa = { workspace = true }
-sp-runtime = { workspace = true, default-features = true }
-beefy-primitives = { workspace = true, default-features = true }
 
 asset-hub-polkadot-runtime = { workspace = true }
 asset-hub-kusama-runtime = { workspace = true }
@@ -50,12 +39,7 @@ runtime-benchmarks = [
 	"encointer-kusama-runtime/runtime-benchmarks",
 	"glutton-kusama-runtime/runtime-benchmarks",
 	"kusama-runtime/runtime-benchmarks",
-	"pallet-staking/runtime-benchmarks",
 	"people-kusama-runtime/runtime-benchmarks",
 	"people-polkadot-runtime/runtime-benchmarks",
-	"polkadot-primitives/runtime-benchmarks",
 	"polkadot-runtime/runtime-benchmarks",
-	"runtime-parachains/runtime-benchmarks",
-	"sp-runtime/runtime-benchmarks",
-	"encointer-kusama-runtime/runtime-benchmarks"
 ]
