+++ relay/kusama/Cargo.toml
@@ -186,7 +186,6 @@ std = [
 	"runtime-parachains/std",
 	"scale-info/std",
 	"serde_json/std",
-	"substrate-wasm-builder",
 	"sp-api/std",
 	"sp-application-crypto/std",
 	"sp-arithmetic/std",
@@ -206,6 +205,7 @@ std = [
 	"sp-tracing/std",
 	"sp-transaction-pool/std",
 	"sp-version/std",
+	"substrate-wasm-builder",
 	"xcm-builder/std",
 	"xcm-executor/std",
 	"xcm/std",
