+++ relay/polkadot/Cargo.toml
@@ -183,7 +183,6 @@ std = [
 	"runtime-parachains/std",
 	"scale-info/std",
 	"serde_json/std",
-	"substrate-wasm-builder",
 	"sp-api/std",
 	"sp-application-crypto/std",
 	"sp-arithmetic/std",
@@ -203,6 +202,7 @@ std = [
 	"sp-tracing/std",
 	"sp-transaction-pool/std",
 	"sp-version/std",
+	"substrate-wasm-builder",
 	"xcm-builder/std",
 	"xcm-executor/std",
 	"xcm/std",
