+++ system-parachains/bridge-hubs/bridge-hub-kusama/Cargo.toml
@@ -139,6 +139,7 @@ std = [
 	"bp-kusama/std",
 	"bp-messages/std",
 	"bp-parachains/std",
+	"bp-polkadot-bulletin/std",
 	"bp-polkadot-core/std",
 	"bp-polkadot/std",
 	"bp-relayers/std",
@@ -219,11 +220,10 @@ std = [
 	"sp-version/std",
 	"substrate-wasm-builder",
 	"system-parachains-constants/std",
+	"tuplex/std",
 	"xcm-builder/std",
 	"xcm-executor/std",
 	"xcm/std",
-	"bp-polkadot-bulletin/std",
-	"tuplex/std"
 ]
 
 runtime-benchmarks = [
